prometheus/pyproject.toml
[project]
name = "prometheus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.28",
    "django-prometheus>=2.4.0",
    "djangorestframework>=3.16.1",
    "numpy>=2.0.2",
]
